# Use itoa instead of Sprinf("%d,"... ) in byte_stream_server_proxy