This should save about 0.2% of CPU for the cache proxy.

using metric.WithLabelValues here didn't help for some reason.

Benchmark results:
```
goos: linux
goarch: amd64
cpu: AMD Ryzen 9 7900X 12-Core Processor
        │    base     │                itoa                │
        │   sec/op    │   sec/op     vs base               │
Read-24   65.22µ ± 4%   61.90µ ± 3%  -5.10% (p=0.025 n=30)

        │     base     │              itoa              │
        │     B/op     │     B/op      vs base          │
Read-24   110.4Ki ± 0%   110.5Ki ± 0%  ~ (p=0.143 n=30)

        │    base    │             itoa             │
        │ allocs/op  │ allocs/op   vs base          │
Read-24   418.0 ± 0%   418.0 ± 0%  ~ (p=0.334 n=30)
```